requirements.txt
aiogram>=3.3.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
typing-extensions>=4.12.0
certifi>=2024.7.4
apscheduler>=3.10.1
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0